% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_alpha.R
\name{add_alpha}
\alias{add_alpha}
\title{Add transparency to a named color}
\usage{
add_alpha(col, alpha)
}
\arguments{
\item{col}{A character string specifying the named color}

\item{alpha}{A numeric value specifying the transparency (0 for fully transparent)}
}
\value{
The output of a \code{rgb} call.
}
\description{
This function converts a named color to RGB and adds an alpha (relative transparency) value.
}
\examples{
transparent_red <- add_alpha("red", 0.3)
}
