% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_modobs2.R
\name{analyse_modobs2}
\alias{analyse_modobs2}
\title{Analyse modelled values versus observed data.}
\usage{
analyse_modobs2(df, mod, obs, type = "points", filnam = NA,
  ylab = "Observed", xlab = "Modeled")
}
\arguments{
\item{mod}{A vector of numeric values representing modelled values.}

\item{obs}{A vector of numeric values representing observed values.}

\item{type}{If \code{"points"}, uses \code{geom_points()}, if \code{"hex"}
uses \code{ggplot2::geom_hex()}, if \code{"heat"} uses adjusted
\code{geom_points()} with color indicating density}

\item{filnam}{A character string specifying the name of the file containing
the plot.}

\item{ylab}{(Optional) A character string for a new y-axis label}

\item{xlab}{(Optional) A character string for a new x-axis label}
}
\description{
Calculates a set of performance statistics and optionally creates plots of modelled
versus observed values.
}
\examples{

}
