% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble_nc.R
\name{as_tibble_nc}
\alias{as_tibble_nc}
\title{Convert NetCDF contents to a data frame.}
\usage{
as_tibble_nc(out)
}
\arguments{
\item{out}{An object returned by the function \link{read_nc_onefile}.}
}
\value{
A tidy data frame with a row for each gridcell. Contains columns
\code{"lon"} (longitude), \code{"lat"} (latitude), and \code{"data"}
(containing all time steps' data for each gridcell.)
}
\description{
Converts the output of \code{"read_nc_onefile()"} into a data frame (tibble)
}
