% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{calc_area}
\alias{calc_area}
\title{Gridcell area}
\usage{
calc_area(lat, dx = 1, dy = 1)
}
\arguments{
\item{lat}{Latitude in degrees}

\item{dx}{Gridcell extent in longitude in degrees. Defaults to 1.}

\item{dy}{Gridcell extent in latitude in degrees. Defaults to 1.}
}
\value{
A numerical value.
}
\description{
Calculates the gridcell area on the spherical Earth with radius 6370499.317638 km.
}
\examples{
area <- calc_area(45, 0.5)

}
