% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSD.color.R
\name{convertcolor}
\alias{convertcolor}
\alias{LSD.convertcolor}
\title{Map R colors to hexadecimal representation}
\usage{
convertcolor(cols, alpha = NULL)
}
\arguments{
\item{cols}{a character vector containing R built-in colors.}

\item{alpha}{alpha value: a two-digit integer between 01 and 99 for color opacity, i.e. appearance of partial or full transparency (usage omitted by default).}
}
\value{
\code{convertcolor} returns a vector containing R built-in colors in hexadecimal representation.
}
\description{
Convert R built-in colors to hexadecimal representation.
}
\examples{
convertcolor(c("red","green","blue"))
}
\seealso{
\code{\link{disco}}, \code{\link{colorpalette}}, \code{\link{demotour}}
}
\author{
Bjoern Schwalb
}
\keyword{color}
