% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutna_headtail.R
\name{cutna_headtail}
\alias{cutna_headtail}
\title{Removes NA values from head and tail}
\usage{
cutna_headtail(vec)
}
\arguments{
\item{vec}{A vector of numeric values.}
}
\value{
A reduced vector, shortened after removing consecutive
NAs from the head and tail of the initial vector (argument \code{vec}).
}
\description{
Removes all consecutive NA values starting from the head
and tail of a vector up to the first non-NA value.
}
\examples{
vec <- c(rep(NA, 5), seq(1,10), NA, seq(12,20), rep(NA,10))
cutna_headtail(vec)
}
