% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutna_headtail_df.R
\name{cutna_headtail_df}
\alias{cutna_headtail_df}
\title{Reduces the rows of a data frame based on the removal of NA values}
\usage{
cutna_headtail_df(df, col, extend = FALSE)
}
\arguments{
\item{df}{A data frame}

\item{col}{A character string specifying the column based on which the
removal of consecutive NA values from head and tail is done.}

\item{extend}{A logical specifying whether NAs at the head or tail of a
data frame's column (\code{col}) should be filled with the first (last)
non-NA value of the respective column. Defaults to \code{FALSE}.}
}
\value{
A reduced data frame, shortened after removing consecutive
NAs from the head and tail of the column (argument \code{col}) of the
initial data frame (argument \code{df}).
}
\description{
Reduces the rows of a data frame based on the removal of NA values from
head and tail of a given column. Considers all consecutive NA values
starting from the head and tail of a vector up to the first non-NA value.
}
\examples{
df <- data.frame( columnname = c(rep(NA, 5), seq(1,10), NA, seq(12,20), rep(NA,10)))
print(cutna_headtail_df(df, "columnname"))
}
