% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_grid.R
\name{df_to_grid}
\alias{df_to_grid}
\title{Converts a data frame to a a grid}
\usage{
df_to_grid(df, varnam, grid = "halfdeg")
}
\arguments{
\item{df}{A data frame. Must have columns named \code{'lon'} and \code{'lat'}
with values shared with longitude and latidue values in the NetCDF file provided
by argument \code{maskfiln}.}

\item{varnam}{A character string specifying the variable name in the data frame.}

\item{grid}{A character string specifying the path of the NetCDF file that is
to be used for defining the array (grid). Defaults to \code{"halfdeg"} (a global
grid with half-degree gridcell size and longitude values ranging from -179.75,
-179.25, ... 179.75, and latitude values from -89.75, -89.25, ... 89.75.}
}
\value{
An 2D matrix
}
\description{
Converts a data frame with longitude and latitude information to a a grid
}
