% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSD.color.R
\name{disco}
\alias{disco}
\alias{LSD.disco}
\alias{display.colorpalette}
\alias{LSD.display.colorpalette}
\title{Disco (DISplays COlorpalettes)}
\usage{
disco(colpal = NULL, nrcol = NULL, alpha = NULL)
}
\arguments{
\item{colpal}{a character vector containing R built-in color names or a name of a \code{LSD} colorpalette as a character string (displays all colorpalettes, if not specified).}

\item{nrcol}{a non-negative integer specifying the number of colors to be used (defaults to length of \code{colpal}, if not specified).}

\item{alpha}{alpha value: a two-digit integer between 01 and 99 for color opacity, i.e. appearance of partial or full transparency (usage omitted by default).}
}
\description{
Displays pre-designed colorpalettes as well as custom-made ones (see \code{colorpalette}).
}
\examples{
disco()
disco("rdbu",10)
}
\seealso{
\code{\link{colorpalette}}, \code{\link{demotour}}
}
\author{
Bjoern Schwalb
}
\keyword{alpha}
\keyword{color,}
