% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nc.R
\name{extract_nc}
\alias{extract_nc}
\title{Extracts point data from a raster file}
\usage{
extract_nc(df, filn, get_time = FALSE)
}
\arguments{
\item{df}{A data frame containing columns \code{lon} and \code{lat}
specifying longitude and latitude values of points for which data is
to be extracted (points organised by rows).}

\item{filn}{A character string specifying the path to a raster file that can be read with
\code{raster::brick} or \code{raster::raster}.}
}
\value{
A data frame of the same number of rows as \code{df} with extracted data
nested in column \code{data}.
}
\description{
Extracts point data from a raster file (e.g. a NetCDF file)
}
