% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_point_nc.R
\name{extract_point_nc}
\alias{extract_point_nc}
\title{Extract point data fron NetCDF file}
\usage{
extract_point_nc(lon, lat, filnam, varnam, verbose = FALSE)
}
\arguments{
\item{lon}{A numeric specifying the longitude of the point for which data is to be read.}

\item{lat}{A numeric specifying the latitude of the point for which data is to be read.}

\item{filnam}{A character string specifying the file name of the NetCDF file.}

\item{varnam}{A character string specifying the variable name of the variable in the NetCDF file.}

\item{verbose}{A boolean specifying whether verbose messages should be printed. Defaults to \code{FALSE}.}
}
\value{
A value or vector of values.
}
\description{
Extracts point data fron NetCDF file, given a geographic location (longitude,
latitude).
}
