% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pointdata.R
\name{extract_pointdata_allsites}
\alias{extract_pointdata_allsites}
\title{Extract point data fron NetCDF file}
\usage{
extract_pointdata_allsites(filename, df_lonlat, time = FALSE)
}
\arguments{
\item{filename}{A numeric specifying the longitude of the point for which data is to be read.}

\item{df_lonlat}{A data frame containing columns \code{"lon"} (longitude in degrees) and \code{"lat"}
(latitude in degrees)}

\item{time}{A logical specifying whether file has a time dimension.}
}
\value{
A data frame with columns \code{"lon"} and \code{"lat"} and a separate column for data.
}
\description{
Extracts point data fron NetCDF file, given a geographic location (longitude,
latitude).
}
