% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_days_since.R
\name{get_days_since}
\alias{get_days_since}
\title{Get days since}
\usage{
get_days_since(
  fromyear,
  len,
  startmoy = 1,
  startdoy = 1,
  by = "months",
  origin = ymd("2001-01-01")
)
}
\arguments{
\item{fromyear}{An integer specifying the start year}

\item{len}{An integer specifying the length of the returned vector}

\item{startmoy}{An integer specifying the month of year (between 1 and 12)
of the first entry of the returned vector. Defaults to 1.}

\item{startdoy}{An integer specifying the day of year (between 1 and 365)
of the first entry of the returned vector. Defaults to 1.}

\item{by}{A character string specifying the "spacing" between dates.
Defaults to \code{"months"}.}
}
\value{
A vector of integers
}
\description{
Constructs a vector of days since a given date
}
