% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_lon.R
\name{integrate_lon}
\alias{integrate_lon}
\title{Integrates a global field}
\usage{
integrate_lon(
  ifil,
  landfil = NA,
  mean_time = TRUE,
  varnam = NA,
  mask = NA,
  method = c("mine", "cdo"),
  ...
)
}
\arguments{
\item{ifil}{Input file name}

\item{landfil}{A separate file providing the land area fraction per gridcell, must have the same resolution and domain as
\code{ifil}}

\item{mean_time}{A logical specifying whether mean across time steps should be taken. Defaults to \code{TRUE}.}

\item{varnam}{(Optional) a variable name}

\item{mask}{(Optional) a mask (2D array) corresponding to spatial dimensions of \code{ifil} containing \code{NA}
where values should be masked out.}

\item{method}{asdf}
}
\value{
A numeric value or vector of numeric values
}
\description{
Calculates the integral of a variable given in per-area units over the entire longitude-latitude domain given in the file
using CDO
}
