% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_lonlat.R
\name{integrate_lonlat}
\alias{integrate_lonlat}
\title{Integrates a global field}
\usage{
integrate_lonlat(ifil, landfil = NA, varnam = NA, ...)
}
\arguments{
\item{ifil}{Input file name}

\item{landfil}{A separate file providing the land area fraction per gridcell, must have the same resolution and domain as
\code{ifil}}

\item{varnam}{(Optional) a variable name}
}
\value{
A numeric value or vector of numeric values
}
\description{
Calculates the integral of a variable given in per-area units over the entire longitude-latitude domain given in the file
using CDO
}
