% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myapprox.R
\name{myapprox}
\alias{myapprox}
\title{Interpolate without much fuss.}
\usage{
myapprox(vec)
}
\arguments{
\item{vec}{A vector of numeric values}
}
\value{
A vector of numeric values
}
\description{
Interpolate without much fuss.
}
\examples{
vec <- c(seq(1,10), NA, seq(12,20))
print(myapprox(vec))
}
