% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myjcolors.R
\name{myjcolors}
\alias{myjcolors}
\title{Colors for a nice palette}
\usage{
myjcolors(ncolors = NA)
}
\arguments{
\item{ncolors}{An integer specifying the number of colors to return.
Retunrs the first \code{ncolors} colors of the palette.}
}
\value{
A named vector of characters specifying the hex codes of the colors
}
\description{
Returns colors from the \code{jcolors::jocolors("default")} palette,
complemented by \code{c("darkgreen", "navyblue", "darkgoldenrod", "orchid")}
}
