% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_selname.R
\name{nc_selname}
\alias{nc_selname}
\title{Subsets by variable}
\usage{
nc_selname(nc, var)
}
\arguments{
\item{nc}{An object returned by \link{read_nc_onefile}}
}
\value{
An nc object.
}
\description{
Subsets an nc object by variable name
}
