% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_selyear.R
\name{nc_selyear}
\alias{nc_selyear}
\title{Subsets by years}
\usage{
nc_selyear(nc, years)
}
\arguments{
\item{nc}{An object returned by \link{read_nc_onefile}}

\item{years}{An integer or vector of integers specifying the years to be selected.}
}
\value{
An nc object.
}
\description{
Subsets an nc object by years
}
