% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_yearsum.R
\name{nc_yearsum}
\alias{nc_yearsum}
\title{Takes mean across time steps}
\usage{
nc_yearsum(ifil, return_ofil = FALSE)
}
\arguments{
\item{ifil}{Input file name}

\item{return_ofil}{A boolean specifying whether the output file name should be returned}
}
\value{
An nc object
}
\description{
Takes mean across time steps using CDO
}
