% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Plots an array as a map}
\usage{
plot_map(arr, lon = NULL, lat = NULL, lev = c(range(c(arr), na.rm =
  TRUE), 10), file = NA, coastlines = TRUE, positive = TRUE,
  toplefttext = NA, toprighttext = NA, color = NA, minval = NA,
  maxval = NA)
}
\arguments{
\item{arr}{A matrix (2D array), interpreted as having longitude in
the first dimension, and latitude in the second.}

\item{lon}{A vector containing the longitude values.}

\item{lat}{A vector containing the longitude values.}

\item{lev}{A vector of length 3 specifying (in that order) the minimum,
level, maximum level, and number of levels used for the color key.
Optional. Defaults to using 10 levels and the minimum and maximum of
non-NA values in \code{arr}.}

\item{file}{(Optional) and output file name for the plot.}

\item{coastlines}{A boolean specifying whether coast lines should be
drawn. Defaults to \code{TRUE}.}

\item{toplefttext}{A character string for annotation in the lop left
corner.}

\item{toprighttext}{A character string for annotation in the lop right
corner.}

\item{color}{A vector of color names (strings) from which colors
for the color key are interpolated. Optional.}

\item{minval}{A numeric value specifying the minimum value used for the
color key.}

\item{maxval}{A numeric value specifying the maximum value used for the
color key.}
}
\value{
The plot
}
\description{
Reads the full content of a NetCDF file, given only the file name
}
