% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map3.R
\name{plot_map3}
\alias{plot_map3}
\title{Plot a nice map with ggplot.}
\usage{
plot_map3(
  obj,
  maxval = NA,
  breaks = NA,
  lonmin = -180,
  lonmax = 180,
  latmin = -90,
  latmax = 90,
  nbin = 10,
  legend_title = waiver(),
  colorscale = viridis::viridis,
  do_reproj = FALSE,
  plot_title = waiver(),
  plot_subtitle = waiver(),
  combine = TRUE,
  varnam = NULL,
  ...
)
}
\arguments{
\item{obj}{An object, either a \code{RasterBrick} (returned from a \code{raster::brick()} function call),
or a list returned from a \code{rbeni::read_nc_onefile()} function call.}

\item{maxval}{A numeric value specifying the maximum value for which the color key is to be extended. Defaults
to \code{NA} (the 99\% quantile of values is used).}

\item{breaks}{A numeric vector specifying the breaks for the color scale. Defaults to \code{NA}, i.e. breaks
are determined automatically based on \code{nbin} and \code{maxval}.}

\item{lonmin}{Left edge (longitude, in degrees), defaults to -180.}

\item{lonmax}{Right edge (longitude, in degrees), defaults to 180.}

\item{latmin}{Lower edge (latitude, in degrees), defaults to -90.}

\item{latmax}{Upper edge (latitude, in degrees), defaults to 90.}

\item{nbin}{An integer specifying the number of bins used for the color key.}

\item{legend_title}{A character string specifying the legend title (annotation above the color key)}

\item{colorscale}{Either function that returns a set of colors or a vector of color names from which to interpolate.
Defaults to \code{virids::viridis}.}

\item{do_reproj}{A boolean specifying whether to re-project the map to Robin projection}

\item{plot_title}{A character string specifying the plot title}

\item{plot_subtitle}{A character string specifying the plot subtitle}

\item{combine}{A boolean specifying whether the map and the colorscale should be combined using cowplot.
Defaults to \code{TRUE}. If \code{FALSE}, a list of elements are retruned, where elements are the ggplot2 plot object
and the coloscale object returned by the call to \link{plot_discrete_cbar}.}

\item{varnam}{If \code{obj} is a rbeni-nc object (returned by \code{read_nc_onefile()}), \code{varnam} must be
provided (a character string specifying the variable name in \code{obj$vars[[varnam]]}).}
}
\value{
A ggplot object for a global map plot.
}
\description{
Returns a cowplot object for a global map plot.
}
