% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_simpl.R
\name{plot_map_simpl}
\alias{plot_map_simpl}
\title{Empty global map}
\usage{
plot_map_simpl(lonmin = -180, lonmax = 180, latmin = -60, latmax = 85)
}
\arguments{
\item{lonmin}{Left edge (longitude, in degrees), defaults to -180.}

\item{lonmax}{Right edge (longitude, in degrees), defaults to 180.}

\item{latmin}{Lower edge (latitude, in degrees), defaults to -90.}

\item{latmax}{Upper edge (latitude, in degrees), defaults to 90.}
}
\value{
A ggplot object for a global map plot.
}
\description{
Returns a ggplot object for an empty global map
}
