% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nc_onefile.R
\name{read_nc_onefile}
\alias{read_nc_onefile}
\title{Reads a NetCDF file}
\usage{
read_nc_onefile(filn)
}
\arguments{
\item{filn}{A character string specifying the full path of the NetCDF
file to be read.}
}
\value{
A list, containing \code{"lon"} (vector of longitudes of
gridcell mid-points), \code{"lat"} (vector of latitudes of gridcell
mid-points), \code{"time"} (vector of lubridate::ymd dates),
\code{"varnams"} (a vector of all variable names as strings), and a
named (nested) list of the data arrays (lon x lat x time) for each
variable.
}
\description{
Reads the full content of a NetCDF file, given only the file name
}
