% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrid_nc.R
\name{regrid_nc}
\alias{regrid_nc}
\title{Plot a nice map with ggplot.}
\usage{
regrid_nc(obj, varname = "", method = "bil", outgrid = "",
  returnobj = FALSE, lonout = NA, latout = NA)
}
\arguments{
\item{obj}{An object, either a \code{RasterBrick} (returned from a \code{raster::brick()} function call),
or a list returned from a \code{rbeni::read_nc_onefile()} function call.}

\item{varname}{A character string specifying the variable name in the NetCDF file that is to be regridded,}

\item{method}{Use \code{method == "bil"} for a bilinear interpolation using \code{cdo remapbil}. Use
\code{"max"} for a regridding by maximum using the raster R package.}

\item{outgrid}{Use \code{outgrid = "halfdeg"} to regrid to half-degree grid (spanning -179.75 to 179.75 in
longitude, and -89.75 to 89.75 in latitude).}

\item{returnobj}{A boolean.}
}
\value{
If argument \code{retrunobj = TRUE}, the regridded object is returned.
}
\description{
Returns a ggplot object for a global map plot.
}
