% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Removes outliers}
\usage{
remove_outliers(vec, coef = 1.5)
}
\arguments{
\item{vec}{A vector of numeric values containing outliers.}

\item{coef}{A numeric value specifying the range of values not considered
outliers (interquartile range times \code{coef}).}
}
\description{
Removes outliers based on inter-quartile range
}
\examples{

}
