% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_nc.R
\name{slice_nc}
\alias{slice_nc}
\title{Slice an nc object}
\usage{
slice_nc(obj, tstep)
}
\arguments{
\item{obj}{a list returned from a \code{rbeni::read_nc_onefile()} function call.}

\item{tstep}{an integer or vector of integers specifying which time step(s) (index starting at 1) to be extracted.}
}
\value{
A list, containing \code{"lon"} (vector of longitudes of
gridcell mid-points), \code{"lat"} (vector of latitudes of gridcell
mid-points), \code{"time"} (vector of lubridate::ymd dates),
\code{"varnams"} (a vector of all variable names as strings), and a
named (nested) list of the data arrays (lon x lat x time) for each
variable.
}
\description{
Extracts a single time step from an rbeni-nc object
}
