% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_modobs2.R
\name{analyse_modobs2}
\alias{analyse_modobs2}
\title{Analyse modelled values versus observed data.}
\usage{
analyse_modobs2(
  df,
  mod,
  obs,
  type = "points",
  filnam = NA,
  xlim = NULL,
  ylim = NULL,
  use_factor = NULL,
  ...
)
}
\arguments{
\item{df}{A data frame containing columns with names corresponding to arguments
\code{mod} and \code{obs}}

\item{mod}{A character string specifying the variable name (column) of the
modelled (simulated) values in data frame \code{df}.}

\item{obs}{A character string specifying the variable name (column) of the
observed values in data frame \code{df}.}

\item{type}{If \code{"points"}, uses \code{geom_points()}, if \code{"hex"}
uses \code{ggplot2::geom_hex()}, if \code{"heat"} uses adjusted
\code{geom_points()} with color indicating density}

\item{filnam}{A character string specifying the name of the file containing
the plot. Defaults to \code{NA} (no file is created).}
}
\description{
Calculates a set of performance statistics and optionally creates plots of modelled
versus observed values.
}
\examples{

}
