% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_modobs2.R
\name{analyse_modobs2}
\alias{analyse_modobs2}
\title{Analyse modelled values versus observed data.}
\usage{
analyse_modobs2(
  df,
  mod,
  obs,
  type = "points",
  filnam = NA,
  relative = FALSE,
  xlim = NULL,
  ylim = NULL,
  use_factor = NULL,
  shortsubtitle = FALSE,
  plot_subtitle = TRUE,
  plot_linmod = TRUE,
  label = FALSE,
  id = NULL,
  nlabels = 1,
  ...
)
}
\arguments{
\item{df}{A data frame containing columns with names corresponding to arguments
\code{mod} and \code{obs}}

\item{mod}{A character string specifying the variable name (column) of the
modelled (simulated) values in data frame \code{df}.}

\item{obs}{A character string specifying the variable name (column) of the
observed values in data frame \code{df}.}

\item{type}{If \code{"points"}, uses \code{geom_points()}, if \code{"hex"}
uses \code{ggplot2::geom_hex()}, if \code{"heat"} uses adjusted
\code{geom_points()} with color indicating density, if \code{"density"} uses
\code{stat_density_2d()} to draws polygos of equal density.}

\item{filnam}{A character string specifying the name of the file containing
the plot. Defaults to \code{NA} (no file is created).}

\item{relative}{A logical specifying whether the relative RMSE and bias (after
division by the mean) is to be showed in the subtitle labels.}

\item{shortsubtitle}{A boolean specifying whether to display only R2, and in
the subtitle metrics.}

\item{plot_subtitle}{A boolean specifying whether to display any metrics. Defaults
to \code{TRUE}.}

\item{plot_linmod}{A boolean specifying whether to display the fitted linear
regression as a red line. Defaults to \code{TRUE}.}

\item{label}{A boolean specifying whether points should be labelled using ggrepel. 
Defaults to \code{FALSE}. Only available for \code{type = "points}. Use argument
\code{nlabels} to specify how many points should be labelled, starting with points
that have the largest residuals from the linear regression fit.}

\item{id}{A character string specifying the column name that identifies the points.
The column's values must be of type integer and is used to label points in case of 
\code{label = TRUE}.}

\item{nlabels}{An integer specifying how many points to be labelled, starting with points
that have the largest residuals from the linear regression fit. Only available 
for \code{type = "points}. Defaults to one.}
}
\description{
Calculates a set of performance statistics and optionally creates plots of modelled
versus observed values.
}
\examples{

}
