% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_noleap_to_ymd.R
\name{conv_noleap_to_ymd}
\alias{conv_noleap_to_ymd}
\title{Convert time vector from a NetCDF file to a ymd object}
\usage{
conv_noleap_to_ymd(time, since, calender.days = "365")
}
\arguments{
\item{time}{A vector with timesteps, which once was the time variable of a
NetCDF file. These timesteps have to be on DAILY basis!}

\item{since}{ymd date object, time starts with.}

\item{calender.days}{Character or numeric. Calendar type. Can be "365" or "360"}
}
\value{
A ymd date vector of dates within NetCDF file, having same length as
NetCDF time dimension.
}
\description{
Converts timesteps from NetCDF file having 365 or 360 days calendar to
lubridate::ymd class timevector. The reason for this function is, that
cannot simply add the timevector to the POSIX date we wish, as this
would lead to complete nonsense data! Therefore, for 365 days calendar
we leave out feb 29th if being leap year.
}
\details{
History: Copied from https://github.com/cran/wux/blob/master/R/ReadNetCdfTimeData.R
}
