% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_grid.R
\name{df_to_grid}
\alias{df_to_grid}
\title{Converts a data frame to a a grid}
\usage{
df_to_grid(df, varnam, lonnam = "lon", latnam = "lat")
}
\arguments{
\item{df}{A data frame. Must have columns named \code{'lon'} and \code{'lat'}
with values shared with longitude and latidue values in the NetCDF file provided
by argument \code{maskfiln}.}

\item{varnam}{A character string specifying the variable name in the data frame.}

\item{lonnam}{A character string specifying the column name containing longitude
values. Defaults to \code{"lon}.}

\item{latnam}{A character string specifying the column name containing longitude
values. Defaults to \code{"lat}.}
}
\value{
A 2D matrix
}
\description{
Converts a data frame with longitude and latitude information to a a grid
}
