% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_to_df.R
\name{nc_to_df}
\alias{nc_to_df}
\title{NetCDF to data frame}
\usage{
nc_to_df(ncfilnam, varnam, do_get_ilon_ilat = FALSE)
}
\arguments{
\item{ncfilnam}{A character string specifying the NetCDF file to be read.}

\item{varnam}{A character string specifying the variable name in the NetCDF file.}

\item{do_get_ilon_ilat}{A boolean specifying whether longitude and latitude
indices are added to the data frame. Defaults to \code{FALSE}.}
}
\value{
A data frame
}
\description{
Reads a NetCDF file into a data frame with longitude and latitude
information
}
