% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nc_onefile.R
\name{read_nc_onefile}
\alias{read_nc_onefile}
\title{Reads a NetCDF file}
\usage{
read_nc_onefile(
  filn,
  varnam = NA,
  date_origin = NA,
  time_is_years = FALSE,
  ignore_time = FALSE,
  check_flip = FALSE
)
}
\arguments{
\item{filn}{A character string specifying the full path of the NetCDF
file to be read.}

\item{varnam}{A character string specifying the variable name of the NetCDF file.
Can also be a vector of character strings. Defaults to \code{NA}, that is: all
available variables are read.}

\item{date_origin}{A character string of format \code{"YYYY-MM-DD"}
specifying the origin, day 0, of the time values provided in the NetCDF file.}

\item{time_is_years}{A logical specifying whether the values provided by dimension
\code{'time'} is years. Defaults to \code{FALSE}.}

\item{ignore_time}{A logical specifying whether file has a time dimension Use this to
ignore if it has a time dimension of length 1 by \code{has_time=TRUE}. Defaults to
\code{FALSE}.}

\item{check_flip}{A logical specifying whether order of latitude values should be checked.}
}
\value{
A list, containing \code{"lon"} (vector of longitudes of
gridcell mid-points), \code{"lat"} (vector of latitudes of gridcell
mid-points), \code{"time"} (vector of lubridate::ymd dates),
\code{"varnams"} (a vector of all variable names as strings), and a
named (nested) list of the data arrays (lon x lat x time) for each
variable.
}
\description{
Reads the full content of a NetCDF file, given only the file name
}
