% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_nc2.R
\name{write_nc2}
\alias{write_nc2}
\title{Write array to a NetCDF file}
\usage{
write_nc2(
  obj,
  varnams = NA,
  filnam_template = NA,
  lonnam_template = "lon",
  latnam_template = "lat",
  lon = NA,
  lat = NA,
  z_dim = NA,
  time = NA,
  make_zdim = FALSE,
  make_tdim = FALSE,
  path = "./out.nc",
  verbose = FALSE,
  lonnam = "lon",
  latnam = "lat",
  zdimnam = "z",
  timenam = "time",
  units_lon = "degrees_east",
  units_lat = "degrees_north",
  units = "NA",
  units_time = "days since 2001-01-01",
  units_zdim = "",
  long_names = "NA",
  missing_value = -9999,
  att_title = "",
  att_history = ""
)
}
\arguments{
\item{obj}{Either an rbeni-nc object or an array of size length(lon) x length(lat) x length(z_dim) x length(time).
The order of lon (1st dim), lat (2nd dim), time (last dim) is mandatory. If it's an rbeni-nc object, then no additional
aruguments are required.}

\item{varnams}{A character string (or vector of strings) specifying the name(s) of the variable(s) that is/are written into the NetCDF output.}

\item{filnam_template}{A character string specifying the file name of the template-NetCDF file from which all dimension information is read
and used for the NetCDF file created here. Defaults to \code{NA}, meaning that all dimension information has to be spedifyed by the user. Othewise all
dimension specification are overridden.}

\item{lonnam_template}{Longitude name in the template file. Required only if \code{!is.na(filnam_template)}.}

\item{latnam_template}{Latitude name in the template file. Required only if \code{!is.na(filnam_template)}.}

\item{lon}{A vector containing the longitudes of the grid cell's mid-points. Defaults to \code{seq(dim(var)[1])}.}

\item{lat}{A vector containing the latitudes of the grid cell's mid-points. Defaults to \code{seq(dim(var)[2])}.}

\item{z_dim}{A vector containing the values of the z-dimension ("z", 3rd dimension). Defaults to NA (no z-dimension is written).}

\item{time}{A vector containing the values of the time dimension. Defaults to NA (no time dimension is written).}

\item{make_zdim}{Write a z-dimension. If var is 2-dimensional, and \code{z_dim==NA}, create a z-dimension of length 1.}

\item{make_tdim}{Write a time dimension. If var is 2-dimensional, and \code{time==NA}, create a time dimension of length 1.}

\item{path}{A character string specifying the file path of the NetCDF file to be written.}

\item{verbose}{A boolean specifying whether messages are to be returned. Defaults to \code{FALSE}.}

\item{lonnam}{Longitude name}

\item{latnam}{Latitude name}

\item{zdimnam}{Z-dimension name}

\item{timenam}{Time dimension name}

\item{units_lon}{Units of longitude}

\item{units_lat}{Units of latitude}

\item{units}{Units of variables.}

\item{units_time}{Units of time dimension.}

\item{units_zdim}{Units of z-dimension.}

\item{long_names}{Long name of \code{var}.}

\item{missing_value}{Missing value for all variables.}

\item{att_title}{Global attribute \code{"Title"}.}

\item{att_history}{Global attribute \code{"History"}.}
}
\value{
The path of the written NetCDF file.
}
\description{
Writes one or multiple arrays with identical dimensions to a NetCDF file,
creating multiple variables each with the same dimension.
}
